moving server script into a module
