+++ packages/libra-rpc-graphql/CHANGELOG.md
@@ -5,7 +5,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
-<!-- ## [Unreleased] -->
+## [Unreleased]
+
+### Added
+
+- added a new `server` module to simplify creating a playground server with `startExpress()` ([#14](https://github.com/Shopify/libra-web-tools/pull/14))
 
 ## [2.1.0]
