+++ packages/libra-rpc-graphql/package.json
@@ -41,7 +41,10 @@
     "@shopify/libra-web-wallet-utils": "^2.0.1",
     "apollo-link": "^1.0.0",
     "apollo-link-schema": "^1.0.0",
+    "apollo-server-express": "^2.0.0",
+    "express": "^4.0.0",
     "graphql": "^14.0.0",
+    "graphql-tag": "^2.0.0",
     "isomorphic-fetch": "^2.0.0"
   },
   "devDependencies": {
@@ -50,9 +53,6 @@
     "@types/ws": "^7.0.0",
     "apollo-cache-inmemory": "^1.0.0",
     "apollo-client": "^2.0.0",
-    "apollo-server-express": "^2.0.0",
-    "express": "^4.0.0",
-    "graphql-tag": "^2.0.0",
     "ts-node-dev": "^1.0.0-pre.44"
   },
   "files": [
