+++ packages/libra-rpc-graphql/src/server/graphql/LibraSeed.ts
@@ -1,0 +1,16 @@
+import gql from 'graphql-tag';
+
+export default gql`
+  mutation LibraSeed {
+    createSeed {
+      phrase
+      key
+      wallet {
+        account {
+          authKey
+          address
+        }
+      }
+    }
+  }
+`;
