+++ packages/libra-rpc-graphql/src/server/graphql/index.ts
@@ -1,0 +1,3 @@
+export {default as libraMintMutation} from './LibraMint';
+export {default as libraSeedMutation} from './LibraSeed';
+export {default as libraTestQuery} from './LibraTestQuery';
