+++ packages/libra-rpc-graphql/src/server/variables.ts
@@ -1,0 +1,10 @@
+// these variables are the result of the LibraSeed mutation
+export const wallet = {
+  phrase:
+    'resource setup claim iron carpet east dose truck dry file olympic february own federal pioneer total candy beef usage heart relax dance library height',
+  key: 'f080ac2748fde3a53318492f8466b200b39afc3b09a496433db6dcca1243953d',
+  secretKey: '81212144b0f8358aaa94ff5f595d5a1274527ce325ff508fe27748e607850463',
+  publicKey: 'dc741284bd42ec0c9250706c4bbe9a651fe8b1ba2cd5b28f9b09f813cb02f3e1',
+  authKey: '239ed09c441bb3307b3c5abc878879581d0a20b6462d3e8dbb21784b1f1314e9',
+  address: '1d0a20b6462d3e8dbb21784b1f1314e9',
+};
